Add Single Player Commands to the archive.
Adds SPC to the archive. Starting with 2.11 in beta 1.7.3, Simo started including an installer for "ease of use". If you are using MultiMC to install this, please use the .zip file and add it as a jar mod for your respective Minecraft version.

The following files are still missing.

==Alpha 1.2.6==
v1.0
v1.1

==Beta 1.2_02==
v1.9
v1.9_01
v2.0
v2.1

==Beta 1.3_01==
v2.3
v2.4
v2.5

==Beta 1.4==
v2.6

==Beta 1.4_01==
v2.7
v2.7_01 / v2.7_1